구분자를 기준으로 분리한 각 숫자의 합을 반환 한다 문자열 계산기 테스트 작성
입력값 검증 테스트 분리
