+++ src/main/kotlin/stringCalculator/StringCalculator.kt
@@ -1,0 +1,4 @@
+package stringCalculator
+
+class StringCalculator {
+}
